Add files via upload
Business Problem: Sales Pipeline Conversion in a SaaS Startup
This assignment is based on a case study on TechnoServe, a fictitious Tech SaaS (software as a service) startup that specialises in offering various software services to small and medium enterprise customers. The products offered by the company are geared towards increasing productivity for its customers. Its software solutions are customised based on the customers’ needs and on how much they are willing to spend on a product.

As TechnoServe is a SaaS provider, its revenue is highly dependent on the number of clients it has. However, the company is facing a pertinent problem that is prevalent in the IT industry today: declining conversions across its sales funnel. This means the company is failing to convert its potential customers into clients.

Your primary objective here would be to find the ideal revenue opportunity in each of these categories and offer recommendations from your final findings.